Wait for a complete timer delegate thread stop
In some cases when timer just stops delegate thread
without waiting for it's complete finishing, there
might be some unexpected data races which can be
avoided if thread is finished completely, especially
if it is going to be restarted like in a periodic
timers. By that reason, thread stopping priority
should be increased to soft stop.